+++ Source/Scene/GlobeSurfaceTileProvider.js
@@ -41,6 +41,7 @@ define([
         './ClippingPlaneCollection',
         './GlobeSurfaceTile',
         './ImageryLayer',
+        './ImagerySplitDirection',
         './QuadtreeTileLoadState',
         './SceneMode',
         './ShadowMode'
@@ -87,6 +88,7 @@ define([
         ClippingPlaneCollection,
         GlobeSurfaceTile,
         ImageryLayer,
+        ImagerySplitDirection,
         QuadtreeTileLoadState,
         SceneMode,
         ShadowMode) {
@@ -929,6 +931,9 @@ define([
             u_scaleAndBias : function() {
                 return this.properties.scaleAndBias;
             },
+            u_splitDirection: function () {
+                return this.properties.splitDirection;
+            },
             u_dayTextureSplit : function() {
                 return this.properties.dayTextureSplit;
             },
@@ -977,6 +982,7 @@ define([
                 dayTextureSaturation : [],
                 dayTextureOneOverGamma : [],
                 dayTextureSplit : [],
+                splitDireciton: ImagerySplitDirection.NONE,
                 dayIntensity : 0.0,
 
                 southAndNorthLatitude : new Cartesian2(),
@@ -1353,6 +1359,9 @@ define([
 
                 ++numberOfDayTextures;
             }
+
+            uniformMapProperties.splitDirection = tileProvider.splitDirection;
+            applySplit = applySplit || tileProvider.splitDirection !== 0.0;
 
             // trim texture array to the used length so we don't end up using old textures
             // which might get destroyed eventually